// SDKs/twilio.com/twilio_taskrouter_v1/1.37.3/typescript/src/sdk/models/operations/listtaskchannel.ts
import { SpeakeasyMetadata, SpeakeasyBase } from "../../../internal/utils";
import * as shared from "../shared";

export const ListTaskChannelServerList = [
	"https://taskrouter.twilio.com",
] as const;


export class ListTaskChannelPathParams extends SpeakeasyBase {
  @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=WorkspaceSid" })
  workspaceSid: string;
}


export class ListTaskChannelQueryParams extends SpeakeasyBase {
  @SpeakeasyMetadata({ data: "queryParam, style=form;explode=true;name=PageSize" })
  pageSize?: number;
}


export class ListTaskChannelSecurity extends SpeakeasyBase {
  @SpeakeasyMetadata({ data: "security, scheme=true;type=http;subtype=basic" })
  accountSidAuthToken: shared.SchemeAccountSidAuthToken;
}


export class ListTaskChannelListTaskChannelResponseMeta extends SpeakeasyBase {
  @SpeakeasyMetadata({ data: "json, name=first_page_url" })
  firstPageUrl?: string;

  @SpeakeasyMetadata({ data: "json, name=key" })
  key?: string;

  @SpeakeasyMetadata({ data: "json, name=next_page_url" })
  nextPageUrl?: string;

  @SpeakeasyMetadata({ data: "json, name=page" })
  page?: number;

  @SpeakeasyMetadata({ data: "json, name=page_size" })
  pageSize?: number;

  @SpeakeasyMetadata({ data: "json, name=previous_page_url" })
  previousPageUrl?: string;

  @SpeakeasyMetadata({ data: "json, name=url" })
  url?: string;
}


export class ListTaskChannelListTaskChannelResponse extends SpeakeasyBase {
  @SpeakeasyMetadata({ data: "json, name=channels", elemType: shared.TaskrouterV1WorkspaceTaskChannel })
  channels?: shared.TaskrouterV1WorkspaceTaskChannel[];

  @SpeakeasyMetadata({ data: "json, name=meta" })
  meta?: ListTaskChannelListTaskChannelResponseMeta;
}


export class ListTaskChannelRequest extends SpeakeasyBase {
  @SpeakeasyMetadata()
  serverUrl?: string;

  @SpeakeasyMetadata()
  pathParams: ListTaskChannelPathParams;

  @SpeakeasyMetadata()
  queryParams: ListTaskChannelQueryParams;

  @SpeakeasyMetadata()
  security: ListTaskChannelSecurity;
}


export class ListTaskChannelResponse extends SpeakeasyBase {
  @SpeakeasyMetadata()
  contentType: string;

  @SpeakeasyMetadata()
  listTaskChannelResponse?: ListTaskChannelListTaskChannelResponse;

  @SpeakeasyMetadata()
  statusCode: number;
}
